+++ js/render-full.js
@@ -1,51 +1,6 @@
-import { createCommentItem, calcHiddenComments } from './comment-layout.js';
-import { COUNT_COMMENTS } from './config.js';
+import { clearExistsСounterComments, createCommentCounterElement, clearExistsComments, renderComments, onClickMoreCommentsButton, buttonLoadMore } from './comment-layout.js';
 
-const closeButton = document.querySelector('.big-picture__cancel');
-const commentsList = document.querySelector('.social__comments');
-const buttonLoadMore = document.querySelector('.comments-loader');
-const counterComments = document.querySelector('.social__comment-count');
-
-const clearExistsComments = () => {
-  document.querySelector('.social__comments').innerHTML = '';
-};
-
-const clearExistsСounterComments = () => {
-  counterComments.innerHTML = '';
-};
-
-const calcAllComments = () => {
-  const countAllComments = commentsList.querySelectorAll('.social__comment');
-  return countAllComments.length;
-};
-
-const calcShowComments = (comments) => {
-  const hiddenComments = calcHiddenComments(comments);
-  const commentsVisible = calcAllComments() - hiddenComments.length;
-  return commentsVisible;
-};
-
-const createCommentCounterElement = () => {
-  const item = document.createElement('span');
-  item.classList.add('comments-count');
-  counterComments.appendChild(item);
-};
-
-const showFiveComments = (hiddenComments, buttonMoreComments) => {
-  hiddenComments.forEach((comment,index) => {
-    if ((index <= COUNT_COMMENTS - 1) && (hiddenComments.length > COUNT_COMMENTS)) {
-      if (buttonMoreComments.classList.contains('hidden')) {
-        buttonMoreComments.classList.remove('hidden');
-      }
-      comment.classList.remove('hidden');
-    } else if (hiddenComments.length <= COUNT_COMMENTS) {
-      buttonMoreComments.classList.add('hidden');
-      comment.classList.remove('hidden');
-    }
-  });
-  const commentCounterElement = document.querySelector('.big-picture .comments-count');
-  commentCounterElement.innerHTML = `${calcShowComments(commentsList)} из <span class="comments-count">${calcAllComments()}</span> комментариев`;
-};
+const buttonClose = document.querySelector('.big-picture__cancel');
 
 const fixBackPage = () => document.querySelector('body').classList.add('modal-open');
 
@@ -54,19 +9,6 @@ const renderPopup = (data) => {
   document.querySelector('.big-picture .big-picture__img img').setAttribute('src', data.url);
   document.querySelector('.big-picture .likes-count').textContent = data.likes;
   document.querySelector('.big-picture .social__caption').textContent = data.description;
-};
-
-const renderComments = (data) => {
-  data.comments
-    .forEach((comment) => {
-      commentsList.appendChild(createCommentItem(comment));
-    });
-  showFiveComments(calcHiddenComments(commentsList), buttonLoadMore);
-};
-
-const onClickMoreCommentsButton = (evt) => {
-  evt.preventDefault();
-  showFiveComments(calcHiddenComments(commentsList), buttonLoadMore);
 };
 
 buttonLoadMore.addEventListener('click', onClickMoreCommentsButton);
@@ -78,7 +20,7 @@ const onDocumentKeydown = (evt) => {
   }
 };
 
-const onClickCloseButton = (evt) => {
+const onClickButtonClose = (evt) => {
   evt.preventDefault();
   closePopup();
 };
@@ -86,7 +28,7 @@ const onClickCloseButton = (evt) => {
 function closePopup () {
   document.querySelector('.big-picture').classList.add('hidden');
   document.querySelector('body').classList.remove('modal-open');
-  closeButton.removeEventListener('click', onClickCloseButton);
+  buttonClose.removeEventListener('click', onClickButtonClose);
   document.removeEventListener('keydown', onDocumentKeydown);
 }
 
@@ -98,7 +40,7 @@ const openFullImage = (data) => {
   clearExistsComments();
   renderComments(data);
   document.addEventListener('keydown', onDocumentKeydown);
-  closeButton.addEventListener('click', onClickCloseButton);
+  buttonClose.addEventListener('click', onClickButtonClose);
 };
 
 export {openFullImage};
